+++ tcex/api/tc/v3/security/owners/owner.py
@@ -9,6 +9,7 @@ from tcex.api.tc.v3.object_abc import ObjectABC
 from tcex.api.tc.v3.object_collection_abc import ObjectCollectionABC
 from tcex.api.tc.v3.security.owners.owner_filter import OwnerFilter
 from tcex.api.tc.v3.security.owners.owner_model import OwnerModel, OwnersModel
+from tcex.pleb.cached_property_filesystem import cached_property_filesystem
 
 
 class Owner(ObjectABC):
@@ -80,6 +81,13 @@ class Owners(ObjectCollectionABC):
     def _api_endpoint(self) -> str:
         """Return the type specific API endpoint."""
         return ApiEndpoints.OWNERS.value
+
+    @cached_property_filesystem(ttl=86400)
+    def cached_dict(self) -> dict[str, dict]:
+        """Return cached data as a dict keyed by name."""
+        return {
+            at.model.name: at.model.dict() for at in self.iterate(base_class=Owner) if at.model.name
+        }
 
     @property
     def filter(self) -> OwnerFilter: