+++ tests/api/tc/v3/security/system_roles/test_system_role_snippets.py
@@ -1,5 +1,8 @@
 """TcEx Framework Module"""
 
+# third-party
+import pytest
+
 # first-party
 from tcex.api.tc.v3.tql.tql_operator import TqlOperator
 from tests.api.tc.v3.v3_helpers import TestV3, V3Helper
@@ -35,3 +38,21 @@ class TestSystemRolesSnippets(TestV3):
         system_role.get()
         print(system_role.model.dict(exclude_none=True))
         # End Snippet
+
+
+    @pytest.mark.parametrize(
+        'name',
+        [
+            ('Administrator'),
+            ('Api User'),
+            ('Read Only User'),
+            ('User'),
+        ],
+    )
+    def test_cached_dict(self, name: str):
+        """Test that the cached data contains the expected name."""
+        # Manual runs of this test case return a runtime of 9 seconds when not cached and
+        # less than 0.1 seconds when cached, so the cache is providing a significant performance
+        # improvement.
+        data = self.tcex.api.tc.v3.system_roles()
+        assert name in data.cached_dict, f'System role {name!r} not found in cache'